Initial commit. I have successfully created a spike for the student list. Currently you can scroll a list of students and if you click them it will return an ID via a Toast.
The next steps are to display the student image data alongside them(which there is already space for), along with a tick as to whether they have a hand up yet.

Further to this, I will be able to swap out these images with video/audio streams once they are fully understood. Once I have got to this point I will be able to reorganise the view to display what I wish it to. Whether this be a grid layout, or some other format.

Adding sizes to the custom row layout seemed to solve the issues of various pieces of text not appearing.